COM-3298 - fix multiple calls + validations
