+++ src/core/composables/learners.js
@@ -29,7 +29,6 @@ export const useLearners = (refresh, isClientInterface, company) => {
   const learnerAlreadyExists = ref(false);
   const traineeAdditionModal = ref(false);
   const newTrainee = ref('');
-  const editedTrainee = ref({ identity: {}, contact: {}, local: {} });
 
   const filteredLearners = computed(() => {
     const formattedString = escapeRegExp(removeDiacritics(searchStr.value));
@@ -45,16 +44,10 @@ export const useLearners = (refresh, isClientInterface, company) => {
       company: { required: requiredIf(!isClientInterface) },
     },
   }));
-  const traineeRules = {
-    newTrainee: { required },
-    editedTrainee: {
-      identity: { lastname: { required } },
-      contact: { phone: { required, frPhoneNumber } },
-    },
-  };
+  const traineeRules = { newTrainee: { required } };
 
   const learnerValidation = useVuelidate(learnerRules, { newLearner });
-  const traineeValidation = useVuelidate(traineeRules, { newTrainee, editedTrainee });
+  const traineeValidation = useVuelidate(traineeRules, { newTrainee });
 
   const goToNextStep = () => {
     firstStep.value = false;
@@ -184,7 +177,6 @@ export const useLearners = (refresh, isClientInterface, company) => {
     learnerAlreadyExists,
     traineeAdditionModal,
     newTrainee,
-    editedTrainee,
     // Computed
     filteredLearners,
     // Validations
